Seperated canvas functions from GfxCont Component
